+++ src/modules/backup.js
@@ -3,9 +3,8 @@ const Link = require("../db/models/link");
 
 module.exports = async ({ count: backupCount }) => {
     const links = await Link.find({}, "-_id -__v");
-    if (!fs.existsSync("backups")) {
-        fs.mkdirSync("backups");
-        console.log("backups!");
+    if (!fs.existsSync("../backups")) {
+        fs.mkdirSync("../backups");
     }
 
     let filename = "backup";
@@ -14,12 +13,12 @@ module.exports = async ({ count: backupCount }) => {
         filename += `-${new Date().toISOString().replace(/:/g, "-")}`;
     }
 
-    fs.writeFileSync(`backups/${filename}.json`, JSON.stringify(links, null, 4));
+    fs.writeFileSync(`../backups/${filename}.json`, JSON.stringify(links, null, 4));
 
     if (backupCount !== -1) {
-        const existingBackups = fs.readdirSync("backups").map((file) => ({
+        const existingBackups = fs.readdirSync("../backups").map((file) => ({
             filename: file,
-            time: fs.statSync(`backups/${file}`).mtime.getTime(),
+            time: fs.statSync(`../backups/${file}`).mtime.getTime(),
         }));
 
         const oldBackups = existingBackups
@@ -28,7 +27,7 @@ module.exports = async ({ count: backupCount }) => {
             .slice(0, backupCount * -1);
 
         oldBackups.forEach((file) => {
-            fs.unlinkSync(`backups/${file}`);
+            fs.unlinkSync(`../backups/${file}`);
         });
     }