+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bug fixes:
   - Fix backup solution to they're saved to /app/backups not /app/src/backups (Introduced in 1.7.0)
   - Importing internal error when importing CSVs from refactor implementing ESLint.
+  - Fix 2fa verification middleware
 - Make users page responsive
 
 ## [1.7.0] - 2023-08-10